requirements.txt
# ═══════════════════════════════════════════════════════════════
# Bybit AI Trading System - Complete Dependencies
# Python 3.10+
# 生成时间: 2025-10-30
# ═══════════════════════════════════════════════════════════════

# ============================================================================
# 核心框架
# ============================================================================

fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-multipart==0.0.6

# ============================================================================
# 数据库
# ============================================================================

sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# ============================================================================
# 认证和安全
# ============================================================================

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ============================================================================
# HTTP客户端
# ============================================================================

requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# ============================================================================
# 数据处理
# ============================================================================

pandas>=2.2.0
numpy>=1.26.0

# ============================================================================
# AI/机器学习
# ============================================================================

openai==1.3.7

# ============================================================================
# 缓存和任务队列（可选）
# ============================================================================

redis==5.0.1
# hiredis==2.2.3  # 需要编译工具，可选
celery==5.3.4

# ============================================================================
# 监控和日志
# ============================================================================

sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
python-json-logger==2.0.7
psutil==5.9.6

# ============================================================================
# 工具库
# ============================================================================

python-dateutil==2.8.2
pytz==2023.3.post1
websockets==12.0
python-socketio==5.10.0

# ============================================================================
# 开发和测试（可选）
# ============================================================================

pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# ============================================================================
# 技术指标（需要系统依赖，暂时注释）
# ============================================================================

# ta-lib==0.4.28  
# 安装说明：
# Ubuntu/Debian: sudo apt-get install ta-lib
# CentOS/RHEL: sudo yum install ta-lib
# macOS: brew install ta-lib
# Windows: 下载预编译包 https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib

# pandas-ta>=0.4.67b0
# 如需使用，请手动安装: pip install pandas-ta

# ═══════════════════════════════════════════════════════════════
# 安装说明
# ═══════════════════════════════════════════════════════════════
#
# 基础安装:
#   pip install -r requirements.txt
#
# 跳过可选依赖:
#   grep -v "^#" requirements.txt | grep -v "redis\|celery\|pytest" | pip install -r /dev/stdin
#
# 更新所有包:
#   pip install -r requirements.txt --upgrade
#
# 检查依赖冲突:
#   pip check
#
# ═══════════════════════════════════════════════════════════════

